requirements.txt
pandas==2.1.1
spacy==3.7.2
sentence-transformers==2.2.2
PyMuPDF==1.23.6
openpyxl==3.1.2
gdown==4.7.1
python-dotenv==1.0.0
openai==0.27.0
urllib3==1.26.15
google-generativeai==0.1.0
numpy==1.23.5
scikit-learn==1.3.2